requirements.txt
Django==4.0.4
Faker==13.11.1
psycopg2==2.9.3
argon2-cffi==21.3.0
django-cors-headers==3.12.0
djangorestframework==3.13.1
django-debug-toolbar==3.4.0
